feat(test): rename to end to end test
