+++ Makefile
@@ -124,9 +124,9 @@ push-docker-uibackend: docker-uibackend ## Build and Push UI Backend Docker imag
 test: ## Run Unit Tests
 	@go test ./...
 
-.PHONY: integration-test
-integration-test: ## Run go integration test against code
-	@cd integration_test && \
+.PHONY: end-to-end-test
+end-to-end-test: ## Run go integration test against code
+	@cd end_to_end_test && \
     go test -v -failfast -test.v -test.paniconexit0 -timeout 2h -ginkgo.v .
 
 .PHONY: clean-ui
