+++ end_to_end_test/default_scan_test.go
@@ -13,25 +13,44 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package integration_test
+package end_to_end_test
 
 import (
 	"fmt"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/openclarity/vmclarity/api/models"
-	"github.com/openclarity/vmclarity/integration_test/helpers"
+	"github.com/openclarity/vmclarity/end_to_end_test/helpers"
 	"github.com/openclarity/vmclarity/pkg/shared/utils"
+	"net/http"
 	"time"
 )
 
-var _ = Describe("Aborting a scan", func() {
+var _ = Describe("Running a default scan (SBOM, vulnerabilities and exploits)", func() {
 
-	Context("which is running", func() {
-		It("should stop successfully", func(ctx SpecContext) {
+	Context("which scans a docker container", func() {
+		It("should finish successfully", func(ctx SpecContext) {
 			By("applying a scan configuration")
 			apiScanConfig, err := client.PostScanConfig(ctx, helpers.GetDefaultScanConfig())
 			Expect(err).NotTo(HaveOccurred())
+
+			By("waiting until grype server is ready")
+			Eventually(func() bool {
+				_, err = http.Get("http://localhost:9991")
+				return err == nil
+			}, time.Second*600).Should(BeTrue())
+
+			By("waiting until trivy server is ready")
+			Eventually(func() bool {
+				_, err = http.Get("http://localhost:9992")
+				return err == nil
+			}, time.Second*600).Should(BeTrue())
+
+			By("waiting until exploit db server is ready")
+			Eventually(func() bool {
+				_, err = http.Get("http://localhost:1326")
+				return err == nil
+			}, time.Second*600).Should(BeTrue())
 
 			By("updating scan configuration to run now")
 			updateScanConfig := helpers.UpdateScanConfigToStartNow(apiScanConfig)
@@ -39,7 +58,7 @@ var _ = Describe("Aborting a scan", func() {
 			Expect(err).NotTo(HaveOccurred())
 
 			By("waiting until scan starts")
-			params := models.GetScansParams{
+			scanParams := models.GetScansParams{
 				Filter: utils.PointerTo(fmt.Sprintf(
 					"scanConfig/id eq '%s' and state ne '%s' and state ne '%s'",
 					*apiScanConfig.Id,
@@ -49,30 +68,24 @@ var _ = Describe("Aborting a scan", func() {
 			}
 			var scans *models.Scans
 			Eventually(func() bool {
-				scans, err = client.GetScans(ctx, params)
+				scans, err = client.GetScans(ctx, scanParams)
 				Expect(err).NotTo(HaveOccurred())
 				return len(*scans.Items) == 1
 			}, helpers.DefaultTimeout, time.Second).Should(BeTrue())
 
-			By("aborting a scan")
-			err = client.PatchScan(ctx, *(*scans.Items)[0].Id, &models.Scan{
-				State: utils.PointerTo(models.ScanStateAborted),
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			By("waiting until scan state changes to aborted")
-			params = models.GetScansParams{
+			By("waiting until scan state changes to done")
+			scanParams = models.GetScansParams{
 				Filter: utils.PointerTo(fmt.Sprintf(
 					"scanConfig/id eq '%s' and state eq '%s'",
 					*apiScanConfig.Id,
-					models.ScanStateAborted,
+					models.ScanStateDone,
 				)),
 			}
 			Eventually(func() bool {
-				scans, err = client.GetScans(ctx, params)
+				scans, err = client.GetScans(ctx, scanParams)
 				Expect(err).NotTo(HaveOccurred())
 				return len(*scans.Items) == 1
-			}, helpers.DefaultTimeout, time.Second).Should(BeTrue())
+			}, time.Second*120, time.Second).Should(BeTrue())
 		})
 	})
 })
