+++ installation/docker/dockercompose.yml
@@ -168,7 +168,7 @@ services:
     image: alpine:3.18.2
     command: ["sleep", "infinity"]
     profiles:
-      - integration-test
+      - end-to-end-test
     labels:
       scanconfig: test
